Initial Push
All assets / Scenes / Folders for start.
Scripts for flock / scientist / individual zombies / random guys.
Player moves with left click. It moves the entire flock.
If any zombie touches random people, they convert and then teleport to a random location around the center of the flock (where the main zombie is located).
If they touch the scientist, it is game over.